move daynight_ratio into astronomy.py
