+++ ecoscope/base/base.py
@@ -1,14 +1,11 @@
 import warnings
 
-import astroplan
-import astropy
 import geopandas as gpd
 import numpy as np
 import pandas as pd
 from pyproj import Geod
 import shapely
 
-from ecoscope.analysis import astronomy
 from ecoscope.base._dataclasses import (
     RelocsCoordinateFilter,
     RelocsDateRangeFilter,
@@ -379,74 +376,6 @@ class Trajectory(EcoDataFrame):
         """
 
         return self["dist_meters"].sum() / self.get_displacement()
-
-    def get_daynight_ratio(self, n_grid_points=150) -> pd.Series:
-        """
-        Parameters
-        ----------
-        n_grid_points : int (optional)
-            The number of grid points on which to search for the horizon
-            crossings of the target over a 24-hour period, default is 150 which
-            yields rise time precisions better than one minute.
-            https://github.com/astropy/astroplan/pull/424
-        Returns
-        -------
-        pd.Series:
-            Daynight ratio for each unique individual subject in the grouby_col column.
-        """
-
-        locations = astronomy.to_EarthLocation(self.geometry.to_crs(crs=self.estimate_utm_crs()).centroid)
-
-        observer = astroplan.Observer(location=locations)
-        is_night_start = observer.is_night(self.segment_start)
-        is_night_end = observer.is_night(self.segment_end)
-
-        # Night -> Night
-        night_distance = self.dist_meters.loc[is_night_start & is_night_end].sum()
-
-        # Day -> Day
-        day_distance = self.dist_meters.loc[~is_night_start & ~is_night_end].sum()
-
-        # Night -> Day
-        night_day_mask = is_night_start & ~is_night_end
-        night_day_df = self.loc[night_day_mask, ["segment_start", "dist_meters", "timespan_seconds"]]
-        i = (
-            pd.to_datetime(
-                astroplan.Observer(location=locations[night_day_mask])
-                .sun_rise_time(
-                    astropy.time.Time(night_day_df.segment_start),
-                    n_grid_points=n_grid_points,
-                    which="next",
-                )
-                .datetime,
-                utc=True,
-            )
-            - night_day_df.segment_start
-        ).dt.total_seconds() / night_day_df.timespan_seconds
-
-        night_distance += (night_day_df.dist_meters * i).sum()
-        day_distance += ((1 - i) * night_day_df.dist_meters).sum()
-
-        # Day -> Night
-        day_night_mask = ~is_night_start & is_night_end
-        day_night_df = self.loc[day_night_mask, ["segment_start", "dist_meters", "timespan_seconds"]]
-        i = (
-            pd.to_datetime(
-                astroplan.Observer(location=locations[day_night_mask])
-                .sun_set_time(
-                    astropy.time.Time(day_night_df.segment_start),
-                    n_grid_points=n_grid_points,
-                    which="next",
-                )
-                .datetime,
-                utc=True,
-            )
-            - day_night_df.segment_start
-        ).dt.total_seconds() / day_night_df.timespan_seconds
-        day_distance += (day_night_df.dist_meters * i).sum()
-        night_distance += ((1 - i) * day_night_df.dist_meters).sum()
-
-        return day_distance / night_distance
 
     @staticmethod
     def _create_multitraj(df):
